requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
pillow==11.1.0
psycopg==3.2.5
psycopg-binary==3.2.5
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
